requirements.txt
requests
python-levenshtein
sqlalchemy
psycopg2
